documentation(REPORT-8206): The blog on Report Parameters Deep Dive into Dynamic and Interactive Reports
