--- webinar/report-parameters -deep-dive-into- dynamic-and- interactive-reports/README.md
@@ -1,7 +1,0 @@
-# Bold Report Webinar - Report Parameters - Deep Dive into dynamic and interactive reports
-
-This will contain the following files
-
-* product-line-sales.rdl and child-report.rdl are the report file which is created during the webinar.
-
-* Report viewer embedded application - created the report viewer app by referring to the getting started document (https://help.boldreports.com/embedded-reporting/javascript-reporting/report-viewer/display-ssrs-rdl-report-in-javascript-application/).
