--- webinar/report-parameters -deep-dive-into- dynamic-and- interactive-reports/Web-ReportViewer-Sample/.vs/ProjectSettings.json
@@ -1,3 +1,0 @@
-{
-  "CurrentProjectSetting": null
-}
