--- webinar/report-parameters -deep-dive-into- dynamic-and- interactive-reports/Web-ReportViewer-Sample/.vs/VSWorkspaceState.json
@@ -1,7 +1,0 @@
-{
-  "ExpandedNodes": [
-    ""
-  ],
-  "SelectedNode": "\\Models",
-  "PreviewInSolutionExplorer": false
-}
