+++ webinar/report-parameters-deep-dive-into-dynamic-and-interactive-reports/README.md
@@ -1,0 +1,14 @@
+# Bold Report Webinar - Report Parameters - Deep Dive into dynamic and interactive reports
+
+This will contain the following files,
+
+* product-line-sales.rdl and child-report.rdl are the report file which is created during the webinar.
+
+* Report Viewer embedded application created by referring to the getting started documentation (https://help.boldreports.com/embedded-reporting/javascript-reporting/report-viewer/display-ssrs-rdl-report-in-javascript-application/).
+
+## How to run the viewer application
+
+* Open the js-report-viewer-app in visual studio 2019.
+* Right click on BoldReportViewer.html and set as startup page.
+* Right click on js-report-viewer-app solution and click on rebuild.
+* Run the application.
