+++ webinar/report-parameters-deep-dive-into-dynamic-and-interactive-reports/js-report-viewer-app/Properties/AssemblyInfo.cs
