+++ webinar/report-parameters-deep-dive-into-dynamic-and-interactive-reports/js-report-viewer-app/Web.config
@@ -18,7 +18,7 @@
       </controls>
     </pages>
     <httpHandlers>
-      <add verb="*" path="captimage.axd" type="Syncfusion.JavaScript.ImageHandler, Syncfusion.EJ, Version=19.2460.0.46, Culture=neutral, PublicKeyToken=3d67ed1f87d44c89" />
+      <add verb="*" path="captimage.axd" type="Syncfusion.JavaScript.ImageHandler, Syncfusion.EJ, Version=18.3460.0.35, Culture=neutral, PublicKeyToken=3d67ed1f87d44c89" />
     </httpHandlers>
     <profile defaultProvider="DefaultProfileProvider">
       <providers>
@@ -39,7 +39,7 @@
       <remove name="OPTIONSVerbHandler" />
       <remove name="TRACEVerbHandler" />
       <add name="ExtensionlessUrlHandler-Integrated-4.0" path="*." verb="*" type="System.Web.Handlers.TransferRequestHandler" preCondition="integratedMode,runtimeVersionv4.0" />
-      <add verb="*" path="captimage.axd" name="syncfusion_generatetools" type="Syncfusion.JavaScript.ImageHandler, Syncfusion.EJ, Version=19.2460.0.46, Culture=neutral, PublicKeyToken=3d67ed1f87d44c89" />
+      <add verb="*" path="captimage.axd" name="syncfusion_generatetools" type="Syncfusion.JavaScript.ImageHandler, Syncfusion.EJ, Version=18.3460.0.35, Culture=neutral, PublicKeyToken=3d67ed1f87d44c89" />
     </handlers>
     <validation validateIntegratedModeConfiguration="false" />
     <modules runAllManagedModulesForAllRequests="true" />
@@ -61,26 +61,6 @@
       <dependentAssembly>
         <assemblyIdentity name="System.Web.WebPages" publicKeyToken="31bf3856ad364e35" />
         <bindingRedirect oldVersion="0.0.0.0-3.0.0.0" newVersion="3.0.0.0" />
-      </dependentAssembly>
-      <dependentAssembly>
-        <assemblyIdentity name="Syncfusion.Pdf.Base" publicKeyToken="3d67ed1f87d44c89" culture="neutral" />
-        <bindingRedirect oldVersion="0.0.0.0-19.2460.0.46" newVersion="19.2460.0.46" />
-      </dependentAssembly>
-      <dependentAssembly>
-        <assemblyIdentity name="Syncfusion.XlsIO.Base" publicKeyToken="3d67ed1f87d44c89" culture="neutral" />
-        <bindingRedirect oldVersion="0.0.0.0-19.2460.0.46" newVersion="19.2460.0.46" />
-      </dependentAssembly>
-      <dependentAssembly>
-        <assemblyIdentity name="Syncfusion.Presentation.Base" publicKeyToken="3d67ed1f87d44c89" culture="neutral" />
-        <bindingRedirect oldVersion="0.0.0.0-19.2460.0.46" newVersion="19.2460.0.46" />
-      </dependentAssembly>
-      <dependentAssembly>
-        <assemblyIdentity name="Syncfusion.DocIO.Base" publicKeyToken="3d67ed1f87d44c89" culture="neutral" />
-        <bindingRedirect oldVersion="0.0.0.0-19.2460.0.46" newVersion="19.2460.0.46" />
-      </dependentAssembly>
-      <dependentAssembly>
-        <assemblyIdentity name="Syncfusion.Licensing" publicKeyToken="632609b4d040f6b4" culture="neutral" />
-        <bindingRedirect oldVersion="0.0.0.0-19.2460.0.46" newVersion="19.2460.0.46" />
       </dependentAssembly>
     </assemblyBinding>
   </runtime>
