+++ webinar/report-parameters-deep-dive-into-dynamic-and-interactive-reports/js-report-viewer-app/js-report-viewer-app.sln
@@ -3,7 +3,7 @@ Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio Version 16
 VisualStudioVersion = 16.0.29806.167
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Web ReportViewer", "Web ReportViewer.csproj", "{AEE06F9F-A352-48F9-8406-E3D7A59BD0E4}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "js-report-viewer-app", "js-report-viewer-app.csproj", "{AEE06F9F-A352-48F9-8406-E3D7A59BD0E4}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
