+++ webinar/report-parameters-deep-dive-into-dynamic-and-interactive-reports/js-report-viewer-app/packages.config
@@ -1,7 +1,7 @@
 ﻿<?xml version="1.0" encoding="utf-8"?>
 <packages>
-  <package id="Bold.Licensing" version="3.1.44" targetFramework="net46" />
-  <package id="BoldReports.Web" version="3.1.44" targetFramework="net46" />
+  <package id="Bold.Licensing" version="3.1.42" targetFramework="net46" />
+  <package id="BoldReports.Web" version="3.1.42" targetFramework="net46" />
   <package id="Microsoft.AspNet.Cors" version="5.2.7" targetFramework="net46" />
   <package id="Microsoft.AspNet.WebApi" version="5.2.7" targetFramework="net46" />
   <package id="Microsoft.AspNet.WebApi.Client" version="5.2.7" targetFramework="net46" />
@@ -10,11 +10,11 @@
   <package id="Microsoft.AspNet.WebApi.WebHost" version="5.2.7" targetFramework="net46" />
   <package id="Microsoft.CodeDom.Providers.DotNetCompilerPlatform" version="3.6.0" targetFramework="net46" />
   <package id="Newtonsoft.Json" version="13.0.1" targetFramework="net46" />
-  <package id="Syncfusion.Compression.Base" version="19.2.0.46" targetFramework="net46" />
-  <package id="Syncfusion.DocIO.AspNet" version="19.2.0.46" targetFramework="net46" />
-  <package id="Syncfusion.Licensing" version="19.2.0.46" targetFramework="net46" />
-  <package id="Syncfusion.OfficeChart.Base" version="19.2.0.46" targetFramework="net46" />
-  <package id="Syncfusion.Pdf.AspNet" version="19.2.0.46" targetFramework="net46" />
-  <package id="Syncfusion.Presentation.AspNet" version="19.2.0.46" targetFramework="net46" />
-  <package id="Syncfusion.XlsIO.AspNet" version="19.2.0.46" targetFramework="net46" />
+  <package id="Syncfusion.Compression.Base" version="18.3.0.35" targetFramework="net46" />
+  <package id="Syncfusion.DocIO.AspNet" version="18.3.0.35" targetFramework="net46" />
+  <package id="Syncfusion.Licensing" version="18.3.0.35" targetFramework="net46" />
+  <package id="Syncfusion.OfficeChart.Base" version="18.3.0.35" targetFramework="net46" />
+  <package id="Syncfusion.Pdf.AspNet" version="18.3.0.35" targetFramework="net46" />
+  <package id="Syncfusion.Presentation.AspNet" version="18.3.0.35" targetFramework="net46" />
+  <package id="Syncfusion.XlsIO.AspNet" version="18.3.0.35" targetFramework="net46" />
 </packages>
